Moved data models package location
